Fold text improvements, multibyte and width.
Multibyte improvement is with replacing strlen() with strdisplaywidth()
which considers multibyte characters to be one space, so that the dashes
are aligned properly.

Fixed a small typo in the documentation by changing references to
"g:votl_modules_load" to "g:vo_modules_load".

Added two options for .vimrc/.vimoutlinerrc with respect to fold text,
the first one is "g:vo_fold_length" which is either "max", in which case
the fold text is padded/trimmed to the window width, or an integer, in
which case the fold text is resized to that width, the default width
being 58, as it was before.
The second option is "g:vo_trim_string" which takes a string with which
the fold text is trimmed to the set length. This option exists because
the user may prefer to use something like "(more)" or "…" instead of the
default ellipsis (...).
However, because I did not know where to document these two options in
the documentation, these features are undocumented, yet.