Don't munge maplocalleader, at least not per default.
Given that 0.3.7 is hopefully getting close, I think it is a time for me
to post here patches we had in Fedora packages.

This one is really old. I had couple of flamewars about this on the
email list with Steve and Noel, but I still believe that this it
redefinition of the default behavior is The Bad Thing™ to do. See

http://thread.gmane.org/gmane.comp.editors.vim.outliner/3524/
http://thread.gmane.org/gmane.comp.editors.vim.outliner/3514/focus=3521

for couple of exchanges on this theme.