Setting the charset of the output file to utf-8
The output of the script now has utf-8 charset tag. I think this is better
because special characters a displayed correctly. Perhaps we can make this
more sophisticate by trying to guess the charset of the original otl file. But
then we have to include some other libraries, what I tried to prevent.