Delete help section 'Post Processors' since it is little more than a pointer to the section on scripts.  The earlier discussion gets the link at the first mention of post processors.
This has the bonus of eliminating an explicit and likely
incorrect path for the scripts directory.

Various minor fixes:
1. Use a single space at the end of a sentence always.

2. Eliminate ~ that ends up displayed.  This was for
section headings that were like this:

============================================
Some Stuff~		   	*votl-some-stuff*

The ~ will change the coloring of the preceding text
ONLY if it is the last  thing on the line.  Since it
is not in these cases, it simply displays literally in
the help viewer.  One pattern, used elsewhere in the help file,
is to use

=============================================
                                 *votl-some-stuff*
Some Stuff~

i.e., use 3 lines instead of 2 so that the left-hand headline
is colored.  But in most cases that I changed there were
already section subheadings that used that style.
So I just deleted the ~.

3. README.md make explicit that only gvim has custom
menus.

4. Remove inappropriate apostrophe in "it's".

5. Fonts are "fixed-width" or "proportional".